+++ R/a_5.R
@@ -7,7 +7,7 @@ NULL
 ##### Dutch adaptation: Same structural pattern with UD labels
 
 a_5 = function(tokens, entities, verb_pos, agent_patient_pos, extract){
-  rule = tquery(token = entities,
+  rule = tquery(OR(token = entities, appos_child = "appos_child"),
                 relation = "conj",
                 label = "Entity", fill = F,
                 parents(pos = c("NOUN", "PROPN", "PRON"),